Added a Park model with attributes Id, Name and Description inside the Model folder. The class is declared Public. There is a ViewModel created with the name ParksViewModel. It inherits the standard ViewModelBase class. This ViewModel contains a RefreshCommand of type AsyncCommand with a getter method. The Parks collection is of type ObservableRangeCollection of model type Park. The refreh method of type Task marks the IsBusy indicator true, clears the above declared collection and load the Parks collection again. The loading of Parks collection is done using a method called LoadParks which adds new objects to the Parks Collection.
The Parks page has a stack layout which consists of a ListView having a background color.
It further contains ItemSource which tells the source of the ListView. We have an Array of type String and this array
contains values like First Item, Second item and so on.

We can also set the template for this ListView using ItemTemplate. It further consists of DataTemplate and ViewCell.

Dependency Injection is implemented to load the parks using Park Service.